TAUR-1313: Moved metric and model helper methods into test_case_base for htmengine
